Fixed Label Overlap
before, the "game over" label and "you win!" label would overlap if the fairy was killed by an enemy before an already fired bullet hit the final enemy. it has been changed so that even if the fairy has been killed by an enemy, if a bullet was fired before hand that would lead to enough points, the player still wins. this means that if the the "game over" label will still show up, but will disappear if the player has enough points.